add new motion sensors (V2)
